requirements.txt
trl
accelerate
datasets
wandb
transformers
torch
openai
python-dotenv
graphviz
nest-asyncio
